pyproject.toml
[tool.poetry]
name = "parcel-level-flood-risk-nyc"
version = "0.1.0"
description = ""
authors = ["Moses McCall <mosesmc52@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "^1.1.1"
apache-airflow = "^3.0.6"
pymongo = "^4.14.1"
jupyterlab = "^4.4.6"
shapely = "^2.1.1"
pyproj = "^3.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
